it works (ish)
It renders to the screen and rotated as expected, however the shading isn't working.
The brightness of each of the triangles is based the dot product of the surface normal and the direction of the light source. This is working and being stored but the recall function is not working as expected.